Dockerfile changes to add required Windows Features
Dockerfile changes to add required Windows Features and to install IIS
URL Rewrite Module.

This is not enought, I can`t install "\WebClient\Microsoft Dynamics NAV
Web Client.msi" because is missing RTC. And to be able to install RTC
the process requires "\Prerequisite Components\Microsoft Visual Studio
2010 Tools For Office Redist\vstor_redist.exe" and this is problem to
install to.

If this is working the next step would be running the following during
the startup script:
New-NAVWebServerInstance -ClientServicesCredentialType NavUserPassword
-ClientServicesPort 7046 -Company 'COMPANY' -DnsIdentity 'NAVSERVER'
-Server 'localhost' -ServerInstance 'NAVSERVICE' -WebServerInstance
'WEBNAV'